# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SimpleWatch API
  version: 0.1.0
...
paths:
...
  /api/v1/rooms/{roomId}/credentials:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                purpose:
                  type: string
                  enum:
                    - voice
                    - whep
              required:
                - purpose
      parameters:
        - schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          in: path
          name: roomId
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  purpose:
                    type: string
                    enum:
                      - voice
                      - whep
                      - whip
                  url:
                    type: string
                    format: uri
                  token:
                    type: string
                    minLength: 1
                  path:
                    type: string
                  expiresAt:
                    type: string
                    format: date-time
                    pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
                required:
                  - purpose
                  - url
                  - token
                  - expiresAt
                additionalProperties: false
...
  /api/v1/internal/mediamtx/auth:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                action:
                  type: string
                  minLength: 1
                path:
                  type: string
                id:
                  nullable: true
                  type: string
              required:
                - action
                - path
              additionalProperties: {}
      responses:
        "204":
          description: Default Response
          content:
            application/json:
              schema:
                type: string
                nullable: true
                enum:
                  - null
...
servers:
  - url: https://watch.simplec.top